new syntax macro (FILE-ENABLE-SQL-READER-SYNTAX) Turns on the SQL reader syntax for the rest of the file. The CL spec says that when finished loading a file the original *readtable* is restored.  clhs COMPILE-FILE
This is because we kept having problems with the read syntax
not being disabled. When clsql goes to enable the syntax it first
checks that it hasn't already done so. Clsql believed it had, but
the lisp had correctly disabled it and so we are left without the
syntax being enabled.