Tests for time printing and retrieving
These tests will create a table and verify that the date inserted is
the date retrieved and that when we update the database the value we
retrieve is the same as the one we inserted.

Some backends support timezones and some do not, also some support
usecs and some dont.

Things the following tests try to prove correct
 * Reading and writing usec and usec-less times
 * reading and writing timezones (Z=utc) when appropriate (eg: postgresql-socket)
 * reading and writing localtimes when appropriate (eg: ODBC)
 * reading and writing through both the oodml and fdml layers

All these tests fail right now. All tests may not be correct yet.