Fixed bugs in iso-timestring regarding printing microseconds
iso-timestring used to output
  YYYY-MM-DD HH:MM:SS,,UUUUUU

What was invalid in this format was the two commas and because 3 usec
would be output as xx,,3 instead of xx,,000003.

It now outputs
  YYYY-MM-DD HH:MM:SS.UUUUUU

Also we only print the amount of fractional seconds necessary.
Different databases support different precisions so this is
intended to ensure that we dont print with a greater
(ie: unparsable)  precision than we read from the DB
   EG 1.123400 sec prints as 1.1234
  and      1.0 sec prints as 1

iso-timestring new implementation tested to be ~2x less memory at
the same speed