Version 5.0.0: International character sets support
First release of CLSQL to formally and consistently support non-ASCII
strings with encoding of external formats for SQL strings.

UFFI version 1.8.1 is higher is required.

This change may introduce some differences in string handling for
people who are using non-ASCII encoded characters.  Thus, because of
the risk of BACKWARD INCOMPATIBILITY, the major version number was
incremented with this release.